# Vectorize IndexOfAnyExcept for four values