@adamsitnik 2022-08-15
@stephentoub Since I currently can't merge #73768, I am going to merge this PR now and address your feedback later (renaming `ComputeIndex` to `ComputeFirstIndex`). Normally I would do it first, but we have limited time ;)